Update jquery.validate.unobtrusive.js
The filter for hidden checkbox items should target actual type of hidden fields rather than :hidden. As the validation will not run correctly when a element is hidden on the page.

Elements can be considered :hidden for several reasons:
They have a CSS display value of none.
They are form elements with type="hidden".
Their width and height are explicitly set to 0.
An ancestor element is hidden, so the element is not shown on the page.